cleaned
commented out debug print statements and added comments to functions and variables. Please change api variables at the top of rendererplugin.tsx in apps/admin/src/plugins/headlessCMS/fields/...renderer.tsx
