+++ apps/admin/src/plugins/headlessCMS/fields/dropdown-select/dropdownFieldPlugin.tsx
@@ -1,22 +1,22 @@
 import React from "react";
 import { CmsEditorFieldTypePlugin } from "@webiny/app-headless-cms/types";
 
-const TextIcon: React.FunctionComponent = () => <i>icon</i>;
+const TextIcon: React.FunctionComponent = () => <i>dropdown select</i>;
 
 const plugin: CmsEditorFieldTypePlugin = {
   type: "cms-editor-field-type",
-  name: "cms-editor-field-type-secret-text",
+  name: "cms-editor-field-type-dropdown",
   field: {
-    type: "secret-text",
-    label: "Secret Text",
-    description: "Store encrypted text into the database.",
+    type: "dropdown",
+    label: "Dropdown Select",
+    description: "Used for selecting mmy and stores into DB.",
     icon: <TextIcon />,
     allowMultipleValues: false,
     allowPredefinedValues: false,
     multipleValuesLabel: "Use as a list of text values",
     createField() {
       return {
-        type: "secret-text",
+        type: "dropdown",
         validation: [],
         renderer: {
           name: ""
